core: migrate data access logic to `@sdk` package for enhanced modularity
- Extracted `profile`, `server`, and `invite` queries/mutations into the new `@sdk` package.
- Updated relevant imports across the codebase to use `@sdk` modules.
- Added path alias `@sdk/*` in `tsconfig.json` for streamlined imports.
