+++ packages/sdk/src/mutations/invite.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query'
+import axios from 'axios'
+
+export const useJoinByInvite = () => {
+  return useMutation({
+    mutationFn: (inviteCode: string) =>
+      axios.post<{ redirectUrl: string }>('/api/socket/servers/invite', { inviteCode }).then((res) => res.data),
+  })
+}
