+++ src/app/(invite)/(routes)/invite/[inviteCode]/invite-page-client.tsx
@@ -8,7 +8,7 @@ import { useTranslations } from 'next-intl'
 import { Spinner } from '@/lib/shared/ui/spinner'
 import { ErrorComponent } from '@/lib/shared/ui/error-component'
 import { Button } from '@/lib/shared/ui/button'
-import { useJoinByInvite } from '@/lib/shared/data-access/server/api-socket'
+import { useJoinByInvite } from '@sdk/mutations/invite'
 import { ERoutes, getSignInRedirectUrl } from '@app-core/routing/routes'
 import { getDesktopInviteUrl } from '@/lib/shared/utils/desktop-deep-link'
 import { getDesktopDownloadUrl } from '@/lib/shared/utils/desktop-download'
