+++ src/lib/server-list/features/server-sidebar.tsx
@@ -18,9 +18,9 @@ import { Spinner } from '@/lib/shared/ui/spinner'
 import { ErrorComponent } from '@/lib/shared/ui/error-component'
 import { UserButton } from '@clerk/nextjs'
 import { useTranslations } from 'next-intl'
-import { useGetServer, useGetServers } from '@/lib/shared/data-access/server/api'
+import { useGetServer, useGetServers } from '@sdk/queries/server'
 import { useSidebarSocket } from '@/lib/shared/data-access/navigation-sidebar/use-sidebar-socket'
-import { useGetProfile } from '@/lib/shared/data-access/user/api'
+import { useGetProfile } from '@sdk/queries/profile'
 import { useRouter } from 'next/navigation'
 import { ERoutes } from '@app-core/routing/routes'
